v3

+++ Ticket.Presentation/ViewModels/TicketDetailReport.cs
@@ -7,8 +7,8 @@ namespace Ticket.Presentation.ViewModels
         public string SubTicketDetail { get; set; }
         public string CreatedBy { get; set; }
         public string WorkingBy { get; set; }
-        public decimal WorkDay { get; set; }
-        public int WorkHour { get; set; }
+        public double WorkDay { get; set; }
+        public double WorkHour { get; set; }
         public string StartDate { get; set; }
     }
 }
